suspense testing
